+++ cmd/server/main.go
@@ -8,8 +8,8 @@ import (
 	"net/http"
 	"os"
 	"os/signal"
-	"pinstack-auth-service/config"
 	token_service "pinstack-auth-service/internal/application/service"
+	"pinstack-auth-service/internal/infrastructure/config"
 	delivery_grpc "pinstack-auth-service/internal/infrastructure/inbound/grpc"
 	auth_grpc "pinstack-auth-service/internal/infrastructure/inbound/grpc/auth"
 	infraLogger "pinstack-auth-service/internal/infrastructure/logger"
